Componentizada la app + browserify + gulp
