+++ index.html
@@ -7,8 +7,6 @@
   </head>
   <body>
     <div class="app-container"></div>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
-    <script src="https://cdn.rawgit.com/visionmedia/page.js/master/page.js"></script>
-    <script src="public/util/routes.js"></script>
+    <script src="public/app.js"></script>
   </body>
 </html>
