+++ lib/app.styl
@@ -11,9 +11,12 @@ test(color= red)
 @import "./battle/styles.styl"
 
 body
-    font-family Arial
-    font-size 16px
-    margin 0
+  font-family Arial
+  font-size 16px
+  margin 0
+
+.hide
+  display none
