+++ lib/index.jade
@@ -6,6 +6,4 @@ html(lang="es")
     link(rel="stylesheet", href="./public/css/app.css")
   body
     .app-container
-    script(src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js")
-    script(src="https://cdn.rawgit.com/visionmedia/page.js/master/page.js")
-    script(src="public/util/routes.js")
+    script(src="public/app.js")
