+++ lib/layout/styles.styl
@@ -16,7 +16,7 @@
     &-status
         test(green)
         width 200px
-        
+
     &-battle
         test(orange)
         flex 1
@@ -34,4 +34,3 @@
         test(gray)
         display flex
         justify-content space-between
-
