--- lib/signin.jade
@@ -1,12 +1,0 @@
-doctype html5
-html(lang="es")
-  head
-    title("Marvel Card Battle")
-    meta(charset="utf-8")
-  body
-    label Inresa tu nombre
-    input.Signin-name-input(type="text", id="firstName", name="firstName", tabindex=1, placeholder="gvilarino, juana, etc...")
-    button.Signin-button(tabindex=2) Ingresar
-    script(src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js")
-    script(src="https://cdn.rawgit.com/visionmedia/page.js/master/page.js")
-    script(src="public/util/routes.js")
