+++ lib/signin/template.jade
@@ -1,0 +1,4 @@
+.signin-container
+  label Ingresa tu nombre
+  input.Signin-name-input(type='text', id='firstName', name='firstName', tabindex=1, placeholder='gvilarino, juan, etc...')
+  button.Signin-button(tabindex=2) Comenzar
