+++ package.json
@@ -1,8 +1,28 @@
 {
+  "dependencies": {
+    "page": "1.6.3",
+    "jquery": "2.1.4"
+  },
+  "devDependencies": {
+    "browserify": "9.0.8",
+    "babelify": "6.1.2",
+    "jadeify": "4.3.0",
+
+    "gulp": "3.9.0",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.0.0",
+
+    "gulp-stylus": "2.0.1",
+    "gulp-concat-css": "2.2.0",
+    "nib": "1.1.0",
+
+    "gulp-uglify": "1.2.0",
+    "gulp-minify-css": "1.1.0"
+
+  },
   "scripts": {
-    "jade": "jade --watch --pretty lib/*.jade --out .",
     "styl": "stylus --use nib --compress --watch lib/app.styl --out public/css",
     "serve": "http-server .",
-    "babel": "babel --watch lib --out-dir public"
+    "build": "browserify -t babelify -t jadeify lib/app.js > public/app.js"
   }
 }
